Revert "refactor operations tests"
This reverts commit 99c71b7b3c51313262ed921492f59e29265fce32.
