+++ tests/conftest.py
@@ -5,16 +5,11 @@ import random
 import string
 import uuid
 import yaml
-import aiohttp_apispec
 from unittest import mock
-from aiohttp_apispec import validation_middleware
-from aiohttp import web
-from pathlib import Path
 
 from app.objects.c_obfuscator import Obfuscator
 from app.utility.base_world import BaseWorld
 from app.service.app_svc import AppService
-from app.service.auth_svc import AuthService
 from app.service.data_svc import DataService
 from app.service.contact_svc import ContactService
 from app.service.event_svc import EventService
@@ -33,12 +28,6 @@ from app.objects.secondclass.c_link import Link
 from app.objects.secondclass.c_fact import Fact
 from app.objects.secondclass.c_relationship import Relationship
 from app.objects.secondclass.c_rule import Rule
-from app.api.v2.responses import json_request_validation_middleware
-from app.api.v2.security import authentication_required_middleware_factory
-from app.api.v2.responses import apispec_request_validation_middleware
-from app.api.v2.handlers.operation_api import OperationApi
-from app.api.rest_api import RestApi
-from app import version
 
 DIR = os.path.dirname(os.path.abspath(__file__))
 CONFIG_DIR = os.path.join(DIR, '..', 'conf')
@@ -254,60 +243,3 @@ def agent_profile():
         )
 
     return _agent_profile
-
-
-@pytest.fixture
-def api_client(loop, aiohttp_client, data_svc):
-    async def initialize():
-        with open(Path(__file__).parents[1] / 'conf' / 'default.yml', 'r') as fle:
-            BaseWorld.apply_config('main', yaml.safe_load(fle))
-        with open(Path(__file__).parents[1] / 'conf' / 'payloads.yml', 'r') as fle:
-            BaseWorld.apply_config('payloads', yaml.safe_load(fle))
-
-        _ = PlanningService()
-        _ = LearningService()
-        auth_svc = AuthService()
-        _ = ContactService()
-        _ = FileSvc()
-
-        def make_app(svcs):
-            app = web.Application(
-                middlewares=[
-                    authentication_required_middleware_factory(services['auth_svc']),
-                    json_request_validation_middleware
-                ]
-            )
-            OperationApi(svcs).add_routes(app)
-            return app
-
-        app_svc = AppService(web.Application())
-        services = app_svc.get_services()
-        os.chdir(str(Path(__file__).parents[1]))
-        app_svc.register_subapp('/api/v2', make_app(services))
-        aiohttp_apispec.setup_aiohttp_apispec(
-            app=app_svc.application,
-            title='CALDERA',
-            version=version.get_version(),
-            swagger_path='/api/docs',
-            url='/api/docs/swagger.json',
-            static_path='/static/swagger'
-        )
-        app_svc.application.middlewares.append(apispec_request_validation_middleware)
-        app_svc.application.middlewares.append(validation_middleware)
-
-        await app_svc.register_contacts()
-        _ = await RestApi(services).enable()
-        await auth_svc.apply(app_svc.application, auth_svc.get_config('users'))
-        await auth_svc.set_login_handlers(services)
-        return app_svc.application
-
-    app = loop.run_until_complete(initialize())
-    return loop.run_until_complete(aiohttp_client(app))
-
-
-@pytest.fixture
-def authorized_cookies(loop, api_client):
-    async def get_cookie():
-        r = await api_client.post('/enter', allow_redirects=False, data=dict(username='admin', password='admin'))
-        return r.cookies
-    return loop.run_until_complete(get_cookie())
